requirements.txt
hydra-core
scikit-learn
matplotlib
seaborn
numpy
wordcloud
scipy
pandas
dvc
pytest
kaggle
great-expectations
apache-airflow==2.7.3
zenml[server]
# feast
pendulum
shyaml
mlflow
# giskard
evidently
Flask
flask-session2
psycopg2-binary
pymystem3
nltk
